refactor : 쿠키 처리를 Cookie 담당 객체에 넘기고 이름을 CookieProvider에서 CookieManager로 수정
